[add/#1] 초기 application 세팅
