+++ app/src/main/java/org/sopt/melon/MelonApplication.kt
@@ -1,0 +1,19 @@
+package org.sopt.melon
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+
+@HiltAndroidApp
+class MelonApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        initTimber()
+    }
+
+    private fun initTimber() {
+        if (BuildConfig.DEBUG) {
+            Timber.plant(Timber.DebugTree())
+        }
+    }
+}
