+++ app/src/main/java/org/sopt/melon/core/designsystem/theme/Theme.kt
@@ -1,4 +1,4 @@
-package org.sopt.melon.ui.theme
+package org.sopt.melon.core.designsystem.theme
 
 import android.app.Activity
 import android.os.Build
