+++ app/src/main/java/org/sopt/melon/presentation/main/MainActivity.kt
@@ -1,0 +1,20 @@
+package org.sopt.melon.presentation.main
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import dagger.hilt.android.AndroidEntryPoint
+import org.sopt.melon.core.designsystem.theme.MELONTheme
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            MELONTheme {
+            }
+        }
+    }
+}
